refactor(spools): modularize ManageFilament component and extract form logic (#112)
